feat: Add wait_for_ip step
Modeled after the VMware plugin's IP wait, this adds a wait
on the availability of the VM's IP and then provides the IP
for use in other steps. If no forwarding is in use and
no IP/host was manually specified, the detected IP will
be used as the default.