feat: Add template_name configuration parameter
When specifying the "name" parameter in a simple HCL configuration,
using a static block definition, it gets correctly defined.

However, when using dynamic blocks the value set for "name" does
not get propagated into the builder configuration structure.

Exposing template_name as a separate parameter, and defaulting it
to "name" if defined, preserves backwards compatibility but also
allows the resulting template volume name to be properly set
when leveraging dynamic blocks in the configuration.